add likes card

+++ src/utils/api.js
@@ -59,20 +59,20 @@ class Api {
 			.then(this._handleResponse)
 	}
 
-	deleteLike(id) {
-		return fetch(`${this._url}cards/${id}/likes`, {
-			headers: this._headers,
-			method: 'DELETE'
-		})
-			.then(this._handleResponse)
-	}
-
-	sendLike(id) {
-		return fetch(`${this._url}cards/${id}/likes`, {
-			headers: this._headers,
-			method: 'PUT'
-		})
-			.then(this._handleResponse)
+	changeLikeCardStatus(id, isLiked) {
+		if(!isLiked){
+			return fetch(`${this._url}cards/${id}/likes`, {
+				headers: this._headers,
+				method: 'PUT'
+			})
+				.then(this._handleResponse)
+		} else {
+			return fetch(`${this._url}cards/${id}/likes`, {
+				headers: this._headers,
+				method: 'DELETE'
+			})
+				.then(this._handleResponse)
+		}
 	}
 
 }
